ci(rust): add Cargo workspace manifest so coverage tooling can run
The OpenCode coverage-evidence job failed with "Rust files changed, but
no root Cargo.toml was found" -- cargo-llvm-cov could not locate a
workspace manifest, so cargo coverage never ran (a tooling/config gap,
not missing tests).

Add a root workspace that contains a new GUI-independent
`bandscope-desktop-core` crate holding the payload contracts, validation
guards, and process helpers that were inlined in the Tauri binary
(src-tauri/src/main.rs), together with the existing unit-test suite. The
Tauri binary now depends on that crate via a path dependency and is
excluded from the workspace, since it links wry/webkit and needs a
bundled frontend, so it is built on its own (see ci.yml).

This lets `cargo llvm-cov` find the root manifest and run the full test
suite from the workspace root on any platform -- no windowing system or
frontend build required.

Verified locally:
- `cargo llvm-cov` runs from the repo root; all 12 tests pass; ~93% line
  coverage on the core crate.
- `cargo check --manifest-path apps/desktop/src-tauri/Cargo.toml
  --locked` still builds the Tauri binary against the extracted crate.
