+++ .gitignore
@@ -8,6 +8,7 @@ __pycache__/
 .mypy_cache/
 .ruff_cache/
 apps/desktop/src-tauri/target/
+/target/
 *.pyc
 *.pyo
 *.egg-info/
