+++ Cargo.toml
@@ -1,0 +1,14 @@
+# Root Cargo workspace manifest.
+#
+# The workspace contains the GUI-independent `bandscope-desktop-core` crate so
+# that coverage tooling (e.g. `cargo llvm-cov`) can locate a root manifest and
+# run the Rust test suite from the workspace root on any platform.
+#
+# The Tauri desktop binary (apps/desktop/src-tauri) is intentionally excluded:
+# it depends on `wry`/webkit and a bundled frontend, so it is built on its own
+# with `--manifest-path apps/desktop/src-tauri/Cargo.toml` (see ci.yml). It
+# consumes `bandscope-desktop-core` as a path dependency.
+[workspace]
+resolver = "2"
+members = ["apps/desktop/core"]
+exclude = ["apps/desktop/src-tauri"]
