+++ apps/desktop/core/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "bandscope-desktop-core"
+version = "0.1.0"
+edition = "2021"
+description = "GUI-independent payload contracts and validation logic for the BandScope desktop app."
+publish = false
+
+[lib]
+name = "bandscope_desktop_core"
+path = "src/lib.rs"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+time = { version = "0.3", features = ["formatting", "macros"] }
+url = "2.5.8"
+
+[dev-dependencies]
+uuid = { version = "1", features = ["v4"] }
