+++ apps/desktop/src-tauri/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 tauri-build = { version = "2", default-features = false, features = [] }
 
 [dependencies]
+bandscope-desktop-core = { path = "../core" }
 rfd = "0.17.2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
